speaker page fix
